services/books/requirements.txt
pydantic==0.32.2
grpcio==1.24.3
protobuf==3.10.0